Renamed display.cpp to Display.cpp so it follows current naming conventions. Fixed some faults that were preventing startup. Removed some debug messages we no longer needed.
