+++ src/engine/Resource.cpp
@@ -21,7 +21,7 @@ ohw::Resource::Resource( const std::string &path, bool persist ) :
 	referencePath( path ),
 	persist( persist ) {
 	u_assert( !path.empty() );
-	DebugMsg( "Created resource, \"%s\"\n", path.c_str() );
+	//DebugMsg( "Created resource, \"%s\"\n", path.c_str() );
 }
 
 ohw::Resource::~Resource() {
